Register long-read post type
This is the post type that will be used for all long-read content.

We could use the extended-cpts package here, but that would run the risk
of conflicts if another plugin (or the theme) was running a different
version of the package. So let's eliminate the risk by just registering
the post type the old-fashioned way.