Add class to generate in-page navigation items
Following on from the previous commit, we need a class that can generate
a list of items that represent H2s within a long-read post to be
navigated to.

This class parses the blocks in the content of a post, and then returns
an array of objects (one for each H2 in the content), where each object
has a 'title' property of the heading text, and an 'id' property of the
id attached to that H2, that we will then be able to use in an anchor
link.

Functionality to ensure WordPress auto-generates IDs for all heading
blocks will need to be activated for this to work correctly, and this
will be enforced in a forthcoming commit.

For: https://trello.com/c/KuFLpg6U/570-%F0%9F%8E%AF-goal-implement-long-read-template-for-lambeth-together-and-enable-gutenberg-for-this