Ensure heading block anchors are auto-generated
In WordPress 5.9, auto-generation of anchors for headings was switched
on. However, in 5.9.1 this was changed to an option you need to
explicitly opt into: WordPress/gutenberg#38780

We need anchors for the heading blocks to always be generated, so that
we can use those anchors for in-page nav.

The suggestion is that this will become opt-out once a UI is available
for these kind of options, probably in WordPress 6.0. So there doesn't
seem to be any significant risk of this functionality being removed in
future, despite it currently being labelled experimental.