Provide static function for getting nav items in theme
This commit adds a new class, `LongReadPlugin\Navigation`, along with a
static method `::getItems()`, that can be called from within a theme to
return all the data required to build the in-page nav.

The method returns an array of objects, in the format:

```
[
  (object) [
    'title' => 'Title of first chapter',
    'url' => 'URL of first chapter'
  ],
  (object) [
    'title' => 'Title of current chapter',
    'url' => null //url is always null for currently viewed chapter
    'subItems' => [
      (object) [
        'title' => 'First in-page nav h2',
        'id' => 'id-of-first-in-page-h2'
      ],
      (object) [
        'title' => 'Second in-page nav h2',
        'id' => 'id-of-second-in-page-h2'
      ]
    ],
    (object) [
      'title' => 'Title of third chapter',
      'url' => 'URL of third chapter'
    ]
    ...
]
```

This should be sufficient to then build out in-page nav using any markup
we like within the theme.

There a couple of extra things to note here:

1. I've chosen to return an empty array from the ::getItems() method if
it is somehow called without the class having already been constructed
(e.g. if there's a template that calls it, but the plugin is
deactivated). This should just result in the nav disappearing, which
seems preferable to a louder error (e.g. an exception being thrown).
2. I've had to suppress two additional Psalm errors in psalm.xml,
because it can't deal with the possibility of properties that are set in
the constructor not being set when you check for them within a method.

The approach of having a static method that depends on an instance of the
class which it is attached to having already been newed-up does feel
slightly counter-intuitive to me, and I'm a bit concerned it's an anti-pattern.
But the only alternative is to have a static method that calls other static
methods, which also feels smelly. Given the static method should only
be called when the plugin is active (and therefore we can be confident
that an instance of the class does exist), I think this is a reasonable
compromise, which keeps the code concise and testable.