Add basic plugin template
This change checks if the option to use the default plugin template has
been checked, and if so loads the plugin template file when viewing a
long read post.

The intention is not that this will be used frequently, but that it
will:

a) provide a quick solution where suitable themes need this added as a
starting point; and
b) serve as a code example of how to use the plugin in a custom
long-read template.