Allow for null type on $useBlockEditor
According to the documentation for the `use_block_editor_for_post_type`
hook, the first parameter should be a boolean: https://developer.wordpress.org/reference/hooks/use_block_editor_for_post_type/

However, when the classic editor plugin is disabled, it seems that this
can actually sometimes be null. This results in a fatal TypeError being
thrown.

Amending this to allow for either a boolean or null will prevent the
error being thrown.