Override user preference for classic editor
Although we've enforced the block editor for the long-read post type in
he `enforceBlockEditor()` method, this only works if the classic editor
is in force globally on the site.

If the classic editor is setup to allow individual editors to choose a
preferred editor, and they've chosen the classic editor, hooking into
`use_block_editor_for_post_type` does not override that selected
preference.

Instead, we have to hook in at the point where the classic editor plugin
retrieves that user option from the database, and always set it to
"block". This is the point in the classic editor code we're hooking
into: https://github.com/WordPress/classic-editor/blob/830fe22808e5b40419560b9119b45ea814ffdd5b/classic-editor.php#L272