Add static function to get title of long-read parent item
In some circumstances, it will make sense to display the title of the
parent long-read item on each long-read page. The expectation is that
this will be done, gov.uk-style, as a span above the h1. But to enable
the most flexibility, let's just create an option that returns the title
of the top ancestor long-read item. (In theory, long-reads should only
ever be one level deep, but we're allowing for greater depths just in
case).

This allows the theme author to display it however they choose, and also
gives us the flexibility for the same function to in future return, for
instance, a custom field that is set on the top-level long-read item
rather than the default title.

This commit also updates the default template to illustrate how this
could be used.