Change in page nav regex to a non-greedy match
This commit changes the regex used to get the id used to generate
the in page navigation links to use a non-greedy match, rather than
a greedy one.

This was motivated by an issue noticed in the NHS England site, where
on pasting in content from a word document the heading anchors were not
reliably operating due to included content. This was determined to be
caused by additional html markup included in the html, which has classes
attached, the greedy match was matching up to the end of the classes in
the additional html, using a non-greedy match resolves this issue safely, by
stopping the regex at the first " encountered, rather than the last.

I have tested this against content with the included formating as per the
article which exhibited the issues, and I have also tested that this works
as expected should the heading being linked to includes "s in the text of
the heading